Move api depot fetch functions in api-depot file
