+++ lib/api-ban.js
@@ -1,7 +1,6 @@
 import HttpError from './http-error'
 
 export const API_BAN_URL = process.env.NEXT_PUBLIC_API_BAN_URL || 'https://plateforme.adresse.data.gouv.fr'
-export const BAN_API_DEPOT = process.env.NEXT_PUBLIC_BAN_API_DEPOT || 'https://plateforme.adresse.data.gouv.fr/api-depot'
 
 export async function _fetch(url) {
   const options = {
@@ -45,16 +44,4 @@ export function getVoieFantoir(voieCode) {
 
 export function getCommune(codeCommune) {
   return _fetch(`${API_BAN_URL}/lookup/${codeCommune}`)
-}
-
-export async function getRevisions(codeCommune) {
-  return _fetch(`${BAN_API_DEPOT}/communes/${codeCommune}/revisions`)
-}
-
-export async function getCurrentRevision(codeCommune) {
-  return _fetch(`${BAN_API_DEPOT}/communes/${codeCommune}/current-revision`)
-}
-
-export function getCurrentBal(codeCommune) {
-  return `${BAN_API_DEPOT}/communes/${codeCommune}/current-revision/files/bal/download`
 }
